Face detection application on camera using Python language
This code will turn on the camera and display the captured image on the screen in real time. It will draw a green rectangle around the detected faces and continue until you press the "q" key. When the "q" key is pressed, the program will close and the camera will be released.